+++ tensorboard/webapp/widgets/data_table/filter_dialog_component.scss
