+++ tensorboard/webapp/widgets/data_table/filter_dialog_module.ts
@@ -14,10 +14,9 @@ limitations under the License.
 ==============================================================================*/
 
 import {NgModule} from '@angular/core';
-import {FilterDialog} from './filter_dialog';
+import {FilterDialog} from './filter_dialog_component';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
-import {MatButtonModule} from '@angular/material/button';
 import {MatCheckboxModule} from '@angular/material/checkbox';
 import {RangeInputModule} from '../range_input/range_input_module';
 import {FilterInputModule} from '../filter_input/filter_input_module';
@@ -26,7 +25,6 @@ import {FilterInputModule} from '../filter_input/filter_input_module';
   declarations: [FilterDialog],
   imports: [
     CommonModule,
-    MatButtonModule,
     MatCheckboxModule,
     FormsModule,
     FilterInputModule,
