+++ tensorboard/webapp/widgets/data_table/filter_dialog_test.ts
@@ -1,8 +1,11 @@
 /* Copyright 2023 The TensorFlow Authors. All Rights Reserved.
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
     http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -11,7 +14,7 @@ limitations under the License.
 ==============================================================================*/
 import {TestBed} from '@angular/core/testing';
 import {DataTableModule} from './data_table_module';
-import {FilterDialog} from './filter_dialog';
+import {FilterDialog} from './filter_dialog_component';
 import {DiscreteFilter, DomainType, IntervalFilter} from './types';
 import {By} from '@angular/platform-browser';
 import {RangeInputComponent} from '../range_input/range_input_component';
